+++ src/font.rs
@@ -5,17 +5,7 @@ use raqote::{DrawOptions, Point, SolidSource};
 
 use crate::DrawTarget;
 
-#[cfg(all(feature = "font-fontkit", feature = "font-fontdue"))]
-std::compile_error!("Multiple font backends are not supported. Choose only a single backend");
-
-#[cfg(feature = "font-fontkit")]
-mod fontkit;
-#[cfg(feature = "font-fontkit")]
-pub type Font = fontkit::Font;
-
-#[cfg(feature = "font-fontdue")]
 mod fdue;
-#[cfg(feature = "font-fontdue")]
 pub type Font = fdue::Font;
 
 pub enum FontColor {