+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased changes
 
+## Changes
+
+- Dropped support for font-kit backend.
+- Use native fontconfig instead of rust-fontconfig.
+
 # 0.2.0 - 2022-10-12
 
 ## Features